requirements.txt
setuptools~=65.5.1
torch~=2.2.2+cu118
numpy~=1.26.4
scikit-image~=0.22.0
albumentations~=1.4.2
scipy~=1.12.0
tqdm~=4.66.2
matplotlib~=3.8.3
pillow~=10.3.0
pandas~=2.2.1